# Make candidate template migrations faster