--- a/packages/@tailwindcss-upgrade/src/codemods/template/is-safe-migration.ts
+++ b/packages/@tailwindcss-upgrade/src/codemods/template/is-safe-migration.ts
@@ -1,3 +1,6 @@
+import { parseCandidate } from '../../../../tailwindcss/src/candidate'
+import type { DesignSystem } from '../../../../tailwindcss/src/design-system'
+
 const QUOTES = ['"', "'", '`']
 const LOGICAL_OPERATORS = ['&&', '||', '?', '===', '==', '!=', '!==', '>', '>=', '<', '<=']
 const CONDITIONAL_TEMPLATE_SYNTAX = [
@@ -12,7 +15,60 @@ const CONDITIONAL_TEMPLATE_SYNTAX = [
 ]
 const NEXT_PLACEHOLDER_PROP = /placeholder=\{?['"]$/
 
-export function isSafeMigration(location: { contents: string; start: number; end: number }) {
+export function isSafeMigration(
+  rawCandidate: string,
+  location: { contents: string; start: number; end: number },
+  designSystem: DesignSystem,
+): boolean {
+  let [candidate] = Array.from(parseCandidate(rawCandidate, designSystem))
+
+  // If we can't parse the candidate, then it's not a candidate at all.
+  if (!candidate) {
+    return false
+  }
+
+  // When we have variants, we can assume that the candidate is safe to migrate
+  // because that requires colons.
+  //
+  // E.g.: `hover:focus:flex`
+  if (candidate.variants.length > 0) {
+    return true
+  }
+
+  // When we have an arbitrary property, the candidate has such a particular
+  // structure it's very likely to be safe.
+  //
+  // E.g.: `[color:red]`
+  if (candidate.kind === 'arbitrary') {
+    return true
+  }
+
+  // A static candidate is very likely safe if it contains a dash.
+  //
+  // E.g.: `items-center`
+  if (candidate.kind === 'static' && candidate.root.includes('-')) {
+    return true
+  }
+
+  // A functional candidate is very likely safe if it contains a value (which
+  // implies a `-`). Or if the root contains a dash.
+  //
+  // E.g.: `bg-red-500`, `bg-position-20`
+  if (
+    (candidate.kind === 'functional' && candidate.value !== null) ||
+    (candidate.kind === 'functional' && candidate.root.includes('-'))
+  ) {
+    return true
+  }
+
+  // If the candidate contains a modifier, it's very likely to be safe because
+  // it implies that it contains a `/`.
+  //
+  // E.g.: `text-sm/7`
+  if (candidate.kind === 'functional' && candidate.modifier) {
+    return true
+  }
+
   let currentLineBeforeCandidate = ''
   for (let i = location.start - 1; i >= 0; i--) {
     let char = location.contents.at(i)!