--- a/packages/@tailwindcss-upgrade/src/codemods/template/is-safe-migration.ts
+++ b/packages/@tailwindcss-upgrade/src/codemods/template/is-safe-migration.ts
@@ -8,6 +8,7 @@ const CONDITIONAL_TEMPLATE_SYNTAX = [
   /v-else-if=['"]$/,
   /v-if=['"]$/,
   /v-show=['"]$/,
+  /(?<!:class)=['"]$/,
 
   // Alpine
   /x-if=['"]$/,