--- a/packages/@tailwindcss-upgrade/src/codemods/template/is-safe-migration.ts
+++ b/packages/@tailwindcss-upgrade/src/codemods/template/is-safe-migration.ts
@@ -9,7 +9,7 @@ const CONDITIONAL_TEMPLATE_SYNTAX = [
   /v-else-if=['"]$/,
   /v-if=['"]$/,
   /v-show=['"]$/,
-  /(?<!:class)=['"]$/,
+  /(?<!:?class)=['"]$/,
 
   // Alpine
   /x-if=['"]$/,