--- a/packages/@tailwindcss-upgrade/src/codemods/template/migrate.ts
+++ b/packages/@tailwindcss-upgrade/src/codemods/template/migrate.ts
@@ -6,6 +6,7 @@ import type { DesignSystem } from '../../../../tailwindcss/src/design-system'
 import { DefaultMap } from '../../../../tailwindcss/src/utils/default-map'
 import { spliceChangesIntoString, type StringChange } from '../../utils/splice-changes-into-string'
 import { extractRawCandidates } from './candidates'
+import { isSafeMigration } from './is-safe-migration'
 import { migrateArbitraryUtilities } from './migrate-arbitrary-utilities'
 import { migrateArbitraryValueToBareValue } from './migrate-arbitrary-value-to-bare-value'
 import { migrateArbitraryVariants } from './migrate-arbitrary-variants'
@@ -59,37 +60,29 @@ export const DEFAULT_MIGRATIONS: Migration[] = [
   migrateOptimizeModifier,
 ]
 
-interface Location {
-  contents: string
-  start: number
-  end: number
-}
-
 let migrateCached = new DefaultMap<
   DesignSystem,
-  DefaultMap<Config | null, DefaultMap<Location | undefined, DefaultMap<string, Promise<string>>>>
+  DefaultMap<Config | null, DefaultMap<string, Promise<string>>>
 >((designSystem) => {
   return new DefaultMap((userConfig) => {
-    return new DefaultMap((location) => {
-      return new DefaultMap(async (rawCandidate) => {
-        let original = rawCandidate
-        for (let migration of DEFAULT_MIGRATIONS) {
-          rawCandidate = await migration(designSystem, userConfig, rawCandidate, location)
-        }
+    return new DefaultMap(async (rawCandidate) => {
+      let original = rawCandidate
+      for (let migration of DEFAULT_MIGRATIONS) {
+        rawCandidate = await migration(designSystem, userConfig, rawCandidate, undefined)
+      }
 
-        // If nothing changed, let's parse it again and re-print it. This will migrate
-        // pretty print candidates to the new format. If it did change, we already had
-        // to re-print it.
-        //
-        // E.g.: `bg-red-500/[var(--my-opacity)]` -> `bg-red-500/(--my-opacity)`
-        if (rawCandidate === original) {
-          for (let candidate of parseCandidate(rawCandidate, designSystem)) {
-            return designSystem.printCandidate(candidate)
-          }
+      // If nothing changed, let's parse it again and re-print it. This will migrate
+      // pretty print candidates to the new format. If it did change, we already had
+      // to re-print it.
+      //
+      // E.g.: `bg-red-500/[var(--my-opacity)]` -> `bg-red-500/(--my-opacity)`
+      if (rawCandidate === original) {
+        for (let candidate of parseCandidate(rawCandidate, designSystem)) {
+          return designSystem.printCandidate(candidate)
         }
+      }
 
-        return rawCandidate
-      })
+      return rawCandidate
     })
   })
 })
@@ -105,7 +98,12 @@ export async function migrateCandidate(
     end: number
   },
 ): Promise<string> {
-  return migrateCached.get(designSystem).get(userConfig).get(location).get(rawCandidate)
+  // Skip this migration if we think that the migration is unsafe
+  if (location && !isSafeMigration(location)) {
+    return rawCandidate
+  }
+
+  return migrateCached.get(designSystem).get(userConfig).get(rawCandidate)
 }
 
 export default async function migrateContents(