--- a/packages/@tailwindcss-upgrade/src/utils/version.ts
+++ b/packages/@tailwindcss-upgrade/src/utils/version.ts
@@ -11,6 +11,15 @@ export function isMajor(version: number) {
   return semver.satisfies(installedTailwindVersion(), `>=${version}.0.0 <${version + 1}.0.0`)
 }
 
+/**
+ * Must be of greater than the current major version including minor and patch.
+ *
+ * E.g.: `isGreaterThan(3)`
+ */
+export function isGreaterThan(version: number) {
+  return semver.gte(installedTailwindVersion(), `${version + 1}.0.0`)
+}
+
 let cache = new DefaultMap((base) => {
   let tailwindVersion = getPackageVersionSync('tailwindcss', base)
   if (!tailwindVersion) throw new Error('Tailwind CSS is not installed')