… CHANGELOG.md
… packages/@tailwindcss-upgrade/src/codemods/template/migrate-canonicalize-candidate.test.ts
… packages/@tailwindcss-upgrade/src/codemods/template/migrate-important.test.ts
… packages/@tailwindcss-upgrade/src/codemods/template/migrate-legacy-classes.test.ts